Restructures components and pages
