+++ src/App.js
@@ -3,10 +3,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { Main } from './Main';
+import Home from './pages/Home';
+import Board from './pages/Board';
+import { path } from 'ramda';
+import { HashRouter, Switch, Route } from 'react-router-dom';
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
-
-import './App.css';
 
 const styles = {
   root: {
@@ -38,7 +39,12 @@ class App extends Component {
       <MuiThemeProvider theme={theme}>
         <div className={classes.root}>
           <CssBaseline />
-          <Main />
+          <HashRouter>
+            <Switch>
+              <Route exact path="/" component={Home} />
+              <Route path="/board/:hash" component={(props) => <Board key={path(['match', 'params', 'hash'], props)} {...props} />} />
+            </Switch>
+          </HashRouter>
         </div>
       </MuiThemeProvider>
     );
